add serializedName in entity & unit test about getting all data

+++ domain/src/main/java/com/example/domain/entity/Data.kt
@@ -27,7 +27,7 @@ data class Data (
     val companyName : String,
     val speakerName : String,
     val speakerLoginYn : String,
-    val reservationUTC : Int,
+    val reservationUTC : Long,
     val reservationYn : String,
     val videoYn : String
 )
